+++ package.json
@@ -23,8 +23,8 @@
     "react": "18.2.0",
     "react-native": "0.72.4",
     "react-native-gesture-handler": "~2.12.0",
-    "react-native-safe-area-context": "4.7.1",
-    "react-native-screens": "~3.24.0"
+    "react-native-safe-area-context": "4.6.3",
+    "react-native-screens": "~3.22.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
